refactor: ChatRoom.jsx 컴포넌트 분리
