+++ src/pages/ChatRoom.css
@@ -1,4 +1,4 @@
-.chat_room_list {
+.chat_message_list {
   height: 70vh;
   background-color: #fff3e9;
   border-radius: 30px;
@@ -10,14 +10,14 @@
   margin: 0 auto;
   overflow: auto;
 }
-.chat_room_list::-webkit-scrollbar {
+.chat_message_list::-webkit-scrollbar {
   width: 8px;
 }
-.chat_room_list::-webkit-scrollbar-track {
+.chat_message_list::-webkit-scrollbar-track {
   background: #ffddbf;
   border-radius: 4px;
 }
-.chat_room_list::-webkit-scrollbar-thumb {
+.chat_message_list::-webkit-scrollbar-thumb {
   background-color: #ffb575;
   border-radius: 4px;
 }
@@ -26,46 +26,46 @@
   justify-content: end;
   text-align: right;
 }
-.chat_room_profile {
+.chat_message_profile {
   display: flex;
   gap: 15px;
   align-items: center;
 }
-.chat_room_profile img {
+.chat_message_profile img {
   width: 70px;
   height: 70px;
   object-fit: cover;
   overflow: hidden;
   border-radius: 50%;
 }
-.chat_room_profile .date {
+.chat_message_profile .date {
   font-size: 13px;
   color: gray;
 }
-.chat_room_profile .message {
+.chat_message_profile .message {
   background-color: white;
   border-radius: 5px;
   padding: 5px 8px;
   width: fit-content;
 }
-.chat_room_profile .my_message {
+.chat_message_profile .my_message {
   background-color: #fecfa9;
   border-radius: 5px;
   padding: 5px 8px;
   width: fit-content;
   margin-left: auto;
 }
-.chat_room_form {
+.chat_message_form {
   max-width: 600px;
   margin: 0 auto;
   display: flex;
 }
-.chat_room_form input {
+.chat_message_form input {
   width: 100%;
   border-top-right-radius: 0;
   border-bottom-right-radius: 0;
 }
-.chat_room_form button {
+.chat_message_form button {
   width: 100px;
   background-color: #ff7800;
   border: none;
